refactor cargo door in common part
