+++ fbw-common/src/wasm/systems/systems/src/hydraulic/mod.rs
@@ -34,6 +34,7 @@ use uom::si::{
 
 pub mod aerodynamic_model;
 pub mod brake_circuit;
+pub mod cargo_doors;
 pub mod electrical_generator;
 pub mod electrical_pump_physics;
 pub mod flap_slat;
